Fix entities with mutual references causing infinite recursion and incorrect equals and hashCode implementations due to all fields being included instead of just primary key fields

+++ BooksOfBliss/src/main/java/com/group05/booksofbliss/model/dao/ListingDAO.java
@@ -9,7 +9,7 @@ import javax.persistence.PersistenceContext;
 import lombok.Getter;
 
 @Stateless
-public class ListingDAO extends AbstractDAO<Listing, Integer> {
+public class ListingDAO extends AbstractDAO<Listing, Long> {
 
     @Getter
     @PersistenceContext(unitName = "bobDB")
